Modified Action and Fixed Bug
- Added convenience parameter to the addUser action (groups, roles,...)
- fixed a bug in the Tokenholder causing unneeded traffic
- fixed formatting in Readme not showing type parameter

+++ src/main/kotlin/de/klg71/keycloakmigration/model/AccessToken.kt
@@ -3,8 +3,8 @@ package de.klg71.keycloakmigration.model
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class AccessToken(@JsonProperty("access_token") val accessToken: String,
-                       @JsonProperty("expires_in") val expiresIn: Int,
-                       @JsonProperty("refresh_expires_in") val refreshExpiresIn: Int,
+                       @JsonProperty("expires_in") val expiresIn: Long,
+                       @JsonProperty("refresh_expires_in") val refreshExpiresIn: Long,
                        @JsonProperty("refresh_token") val refreshToken: String,
                        @JsonProperty("token_type") val tokenType: String,
                        @JsonProperty("not-before-policy") val notBeforePolicy: Int,
